+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
+## 3.2.9 (2024-02-10)
+### Fix
+- remove pandas as dependency for telegram reports
+- show when config file and filter file have been saved
+- better logging information
+## 3.2.8 (2023-01-24)
+### Fix
+- removed the language check
 ## 3.2.7 (2023-09-30)
 ### Fix
 - using the monkey approach until this bug is fixed https://github.com/openatx/atx-agent/pull/111